Start moving types into own files
