+++ src/types/index.ts
@@ -1,1 +1,3 @@
 export * from "./types";
+export * from "./queryStrings";
+export * from "./themeConfig";
