+++ src/types/types.ts
@@ -289,80 +289,3 @@ export interface TDictionary<T> {
 }
 
 export type TOrganisationDictionary = TDictionary<TOrganisation>;
-
-export type TQueryStrings = {
-  // Core
-  query_string: string;
-  exact_match: string;
-  region: string;
-  country: string;
-  category: string;
-  year_range: string;
-  sort_field: string;
-  sort_order: string;
-  offset: string;
-  active_continuation_token: string;
-  continuation_tokens: string;
-  // Multilater Climate Funds (MCF)
-  fund: string;
-  status: string;
-  implementing_agency: string;
-};
-
-export type TDocumentCategory = "All" | "Laws" | "Policies" | "UNFCCC" | "Litigation" | "MCF";
-
-// Theme configuration types
-export type TLabelVariation = {
-  category: string[];
-  label: string;
-  key: string;
-};
-
-export type TThemeConfigOption = {
-  label: string;
-  slug: string;
-  value?: string[];
-  category?: string[];
-  corporaKey?: string;
-  alias?: string;
-};
-
-type TThemeConfigCategory = {
-  label: string;
-  options: TThemeConfigOption[];
-};
-
-export type TThemeConfigFilter = {
-  label: string;
-  taxonomyKey: string;
-  apiMetaDataKey?: string;
-  type: string;
-  category: string[];
-  startOpen?: "true" | "false";
-  options?: TThemeConfigOption[];
-  showFade?: "true" | "false";
-  dependantFilterKey?: string;
-  corporaKey?: string;
-  quickSearch?: string;
-};
-
-type TThemeLink = {
-  key: string;
-  url: string;
-};
-
-type TThemeMetadata = {
-  key: string;
-  title: string;
-  description: string;
-};
-
-export type TThemeConfig = {
-  defaultCorpora?: string[];
-  categories?: TThemeConfigCategory;
-  filters: TThemeConfigFilter[];
-  labelVariations: TLabelVariation[];
-  links: TThemeLink[];
-  metadata: TThemeMetadata[];
-  documentCategories: TDocumentCategory[];
-};
